modify reservation model and adapt tests

+++ backend/src/main/java/org/example/backend/model/Reservation.java
@@ -9,5 +9,5 @@ public record Reservation(String id,
                           Presentation presentation,
                           int amountOfSeats,
                           List<SeatPosition> seatPositions,
-                          double price) {
+                          int price) {
 }
