+++ backend/src/main/java/org/example/backend/model/ReservationDto.java
@@ -5,5 +5,5 @@ import java.util.List;
 public record ReservationDto(Presentation presentation,
                              int amountOfSeats,
                              List<SeatPosition> seatPositions,
-                             double price) {
+                             int price) {
 }
